+++ app/src/main/java/org/horaapps/leafpic/fragments/RvMediaFragment.java
@@ -4,13 +4,19 @@ import android.content.Context;
 import android.content.DialogInterface;
 import android.content.Intent;
 import android.content.res.Configuration;
+import android.graphics.Bitmap;
+import android.graphics.PorterDuff;
+import android.graphics.PorterDuffColorFilter;
 import android.net.Uri;
+import android.os.AsyncTask;
 import android.os.Bundle;
 import android.support.annotation.Nullable;
 import android.support.v4.widget.SwipeRefreshLayout;
 import android.support.v7.app.AlertDialog;
+import android.support.v7.widget.CardView;
 import android.support.v7.widget.GridLayoutManager;
 import android.support.v7.widget.RecyclerView;
+import android.support.v7.widget.SwitchCompat;
 import android.util.Log;
 import android.view.LayoutInflater;
 import android.view.Menu;
@@ -20,8 +26,16 @@ import android.view.View;
 import android.view.ViewGroup;
 import android.view.animation.OvershootInterpolator;
 import android.widget.EditText;
+import android.widget.LinearLayout;
+import android.widget.RadioButton;
+import android.widget.RadioGroup;
+import android.widget.ScrollView;
+import android.widget.SeekBar;
+import android.widget.TextView;
+import android.widget.Toast;
 
 import com.mikepenz.google_material_typeface_library.GoogleMaterial;
+import com.mikepenz.iconics.view.IconicsImageView;
 import com.orhanobut.hawk.Hawk;
 
 import org.horaapps.leafpic.R;
@@ -40,6 +54,7 @@ import org.horaapps.leafpic.data.filter.MediaFilter;
 import org.horaapps.leafpic.data.provider.CPHelper;
 import org.horaapps.leafpic.data.sort.SortingMode;
 import org.horaapps.leafpic.data.sort.SortingOrder;
+import org.horaapps.leafpic.util.Affix;
 import org.horaapps.leafpic.util.AlertDialogsHelper;
 import org.horaapps.leafpic.util.Measure;
 import org.horaapps.leafpic.util.StringUtils;
@@ -421,6 +436,193 @@ public class RvMediaFragment extends BaseFragment {
                                     adapter.clearSelected();
                                 });
                 return true;
+
+            //region Affix
+            // TODO: 11/21/16 move away from here
+            case R.id.affix:
+
+                //region Async MediaAffix
+                class affixMedia extends AsyncTask<Affix.Options, Integer, Void> {
+                    private AlertDialog dialog;
+
+                    @Override
+                    protected void onPreExecute() {
+                        super.onPreExecute();
+                        dialog = AlertDialogsHelper.getProgressDialog((ThemedActivity) getActivity(), getString(R.string.affix), getString(R.string.affix_text));
+                        dialog.show();
+                    }
+
+                    @Override
+                    protected Void doInBackground(Affix.Options... arg0) {
+                        ArrayList<Bitmap> bitmapArray = new ArrayList<Bitmap>();
+                       for (int i = 0; i<adapter.getSelectedCount(); i++) {
+                            if(!adapter.getSelected().get(i).isVideo())
+                                bitmapArray.add(adapter.getSelected().get(i).getBitmap());
+                       }
+
+                        if (bitmapArray.size() > 1)
+                            Affix.AffixBitmapList(getActivity(), bitmapArray, arg0[0]);
+                        else getActivity().runOnUiThread(new Runnable() {
+                            @Override
+                            public void run() {
+                                Toast.makeText(getContext(), R.string.affix_error, Toast.LENGTH_SHORT).show();
+                            }
+                        });
+                        return null;
+                    }
+
+                    @Override
+                    protected void onPostExecute(Void result) {
+                        adapter.clearSelected();
+                        dialog.dismiss();
+                    }
+                }
+                //endregion
+
+                final AlertDialog.Builder builder = new AlertDialog.Builder((ThemedActivity) getActivity(), getDialogStyle());
+                final View dialogLayout = getLayoutInflater().inflate(R.layout.dialog_affix, null);
+
+                dialogLayout.findViewById(R.id.affix_title).setBackgroundColor(getPrimaryColor());
+                ((CardView) dialogLayout.findViewById(R.id.affix_card)).setCardBackgroundColor(getCardBackgroundColor());
+
+                //ITEMS
+                final SwitchCompat swVertical = (SwitchCompat) dialogLayout.findViewById(R.id.affix_vertical_switch);
+                final SwitchCompat swSaveHere = (SwitchCompat) dialogLayout.findViewById(R.id.save_here_switch);
+
+                final LinearLayout llSwVertical = (LinearLayout) dialogLayout.findViewById(R.id.ll_affix_vertical);
+                final LinearLayout llSwSaveHere = (LinearLayout) dialogLayout.findViewById(R.id.ll_affix_save_here);
+
+                final RadioGroup radioFormatGroup = (RadioGroup) dialogLayout.findViewById(R.id.radio_format);
+
+                final TextView txtQuality = (TextView) dialogLayout.findViewById(R.id.affix_quality_title);
+                final SeekBar seekQuality = (SeekBar) dialogLayout.findViewById(R.id.seek_bar_quality);
+
+                //region Example
+                final LinearLayout llExample = (LinearLayout) dialogLayout.findViewById(R.id.affix_example);
+                llExample.setBackgroundColor(getBackgroundColor());
+                llExample.setVisibility(Hawk.get("show_tips", true) ? View.VISIBLE : View.GONE);
+                final LinearLayout llExampleH = (LinearLayout) dialogLayout.findViewById(R.id.affix_example_horizontal);
+                //llExampleH.setBackgroundColor(getCardBackgroundColor());
+                final LinearLayout llExampleV = (LinearLayout) dialogLayout.findViewById(R.id.affix_example_vertical);
+                //llExampleV.setBackgroundColor(getCardBackgroundColor());
+
+
+                //endregion
+
+                //region THEME STUFF
+                getThemeHelper().setScrollViewColor((ScrollView) dialogLayout.findViewById(R.id.affix_scrollView));
+
+                /** TextViews **/
+                int color = getTextColor();
+                ((TextView) dialogLayout.findViewById(R.id.affix_vertical_title)).setTextColor(color);
+                ((TextView) dialogLayout.findViewById(R.id.compression_settings_title)).setTextColor(color);
+                ((TextView) dialogLayout.findViewById(R.id.save_here_title)).setTextColor(color);
+
+                //Example Stuff
+                ((TextView) dialogLayout.findViewById(R.id.affix_example_horizontal_txt1)).setTextColor(color);
+                ((TextView) dialogLayout.findViewById(R.id.affix_example_horizontal_txt2)).setTextColor(color);
+                ((TextView) dialogLayout.findViewById(R.id.affix_example_vertical_txt1)).setTextColor(color);
+                ((TextView) dialogLayout.findViewById(R.id.affix_example_vertical_txt2)).setTextColor(color);
+
+
+                /** Sub TextViews **/
+                color = getThemeHelper().getSubTextColor();
+                ((TextView) dialogLayout.findViewById(R.id.save_here_sub)).setTextColor(color);
+                ((TextView) dialogLayout.findViewById(R.id.affix_vertical_sub)).setTextColor(color);
+                ((TextView) dialogLayout.findViewById(R.id.affix_format_sub)).setTextColor(color);
+                txtQuality.setTextColor(color);
+
+                /** Icons **/
+                color = getIconColor();
+                ((IconicsImageView) dialogLayout.findViewById(R.id.affix_quality_icon)).setColor(color);
+                ((IconicsImageView) dialogLayout.findViewById(R.id.affix_format_icon)).setColor(color);
+                ((IconicsImageView) dialogLayout.findViewById(R.id.affix_vertical_icon)).setColor(color);
+                ((IconicsImageView) dialogLayout.findViewById(R.id.save_here_icon)).setColor(color);
+
+                //Example bg
+                color = getCardBackgroundColor();
+                ((TextView) dialogLayout.findViewById(R.id.affix_example_horizontal_txt1)).setBackgroundColor(color);
+                ((TextView) dialogLayout.findViewById(R.id.affix_example_horizontal_txt2)).setBackgroundColor(color);
+                ((TextView) dialogLayout.findViewById(R.id.affix_example_vertical_txt1)).setBackgroundColor(color);
+                ((TextView) dialogLayout.findViewById(R.id.affix_example_vertical_txt2)).setBackgroundColor(color);
+
+                seekQuality.getProgressDrawable().setColorFilter(new PorterDuffColorFilter(getAccentColor(), PorterDuff.Mode.SRC_IN));
+                seekQuality.getThumb().setColorFilter(new PorterDuffColorFilter(getAccentColor(), PorterDuff.Mode.SRC_IN));
+
+                getThemeHelper().themeRadioButton((RadioButton) dialogLayout.findViewById(R.id.radio_jpeg));
+                getThemeHelper().themeRadioButton((RadioButton) dialogLayout.findViewById(R.id.radio_png));
+                getThemeHelper().themeRadioButton((RadioButton) dialogLayout.findViewById(R.id.radio_webp));
+                getThemeHelper().setSwitchCompactColor( swSaveHere, getAccentColor());
+                getThemeHelper().setSwitchCompactColor( swVertical, getAccentColor());
+                //#endregion
+
+                seekQuality.setOnSeekBarChangeListener(new SeekBar.OnSeekBarChangeListener() {
+                    @Override
+                    public void onProgressChanged(SeekBar seekBar, int progress, boolean fromUser) {
+                        txtQuality.setText(StringUtils.html(String.format(Locale.getDefault(), "%s <b>%d</b>", getString(R.string.quality), progress)));
+                    }
+
+                    @Override
+                    public void onStartTrackingTouch(SeekBar seekBar) {
+
+                    }
+
+                    @Override
+                    public void onStopTrackingTouch(SeekBar seekBar) {
+
+                    }
+                });
+                seekQuality.setProgress(50);
+
+                swVertical.setClickable(false);
+                llSwVertical.setOnClickListener(new View.OnClickListener() {
+                    @Override
+                    public void onClick(View v) {
+                        swVertical.setChecked(!swVertical.isChecked());
+                        getThemeHelper().setSwitchCompactColor(swVertical, getAccentColor());
+                        llExampleH.setVisibility(swVertical.isChecked() ? View.GONE : View.VISIBLE);
+                        llExampleV.setVisibility(swVertical.isChecked() ? View.VISIBLE : View.GONE);
+                    }
+                });
+
+                swSaveHere.setClickable(false);
+                llSwSaveHere.setOnClickListener(new View.OnClickListener() {
+                    @Override
+                    public void onClick(View v) {
+                        swSaveHere.setChecked(!swSaveHere.isChecked());
+                        getThemeHelper().setSwitchCompactColor(swSaveHere, getAccentColor());
+                    }
+                });
+
+                builder.setView(dialogLayout);
+                builder.setPositiveButton(this.getString(R.string.ok_action).toUpperCase(), new DialogInterface.OnClickListener() {
+                    public void onClick(DialogInterface dialog, int id) {
+                        Bitmap.CompressFormat compressFormat;
+                        switch (radioFormatGroup.getCheckedRadioButtonId()) {
+                            case R.id.radio_jpeg:
+                            default:
+                                compressFormat = Bitmap.CompressFormat.JPEG;
+                                break;
+                            case R.id.radio_png:
+                                compressFormat = Bitmap.CompressFormat.PNG;
+                                break;
+                            case R.id.radio_webp:
+                                compressFormat = Bitmap.CompressFormat.WEBP;
+                                break;
+                        }
+
+                        Affix.Options options = new Affix.Options(
+                                swSaveHere.isChecked() ? adapter.getFirstSelected().getPath() : Affix.getDefaultDirectoryPath(),
+                                compressFormat,
+                                seekQuality.getProgress(),
+                                swVertical.isChecked());
+                        new affixMedia().execute(options);
+                    }
+                });
+                builder.setNegativeButton(this.getString(R.string.cancel).toUpperCase(), null);
+                builder.show();
+                return true;
+            //endregion
         }
 
         return super.onOptionsItemSelected(item);
